Modify ScrollDurationManger.java
